fix: audio analyzer hook
